+++ src/components/spartan/SCardBrand/index.ts
@@ -1,0 +1,3 @@
+export { default as SMethodIcon } from './SCardBrand.vue';
+
+export type { TCardBrandProps } from './types';