+++ src/components/spartan/index.ts
@@ -43,3 +43,5 @@ export * from './SInputMask';
 export * from './SInputMaskBlock';
 export * from './SInputDate';
 export * from './SInputDateBlock';
+export * from './SCardBrand';
+export * from './SPageTitle';